backend/requirements.txt
alembic==1.18.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.22
requests==2.32.5
SQLAlchemy==2.0.46
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
